[Update] Change the structure AnimationState and AnimationController to become a generic class. The reason for this is that we have different [GUI::DataStructure] i.e 7 Data Structure to aid our visualization, and when we create a new AnimationState object, we must define the GUI class of the DataStructure that we want to visualize. Therefore, we must have different AnimationState<GUI::DataStructure> class corresponse to different GUI::DataStructure. Also, I have update the AnimationController class of each scene (State).
So, to wrap up, every Data Structure will have its
own GUI visualization. Our AnimationController
will be responsible for handling user interaction
(such as changing animation speed, run all and
reset, etc). When we click an action, an array of
animation will be generated and rendered. One
animation is one
AnimationState<GUI::[DataStructure]>, storing
basic infos that aid for the animation (i.e
duration, highlighted line of code, etc) and a
copy of the GUI of the corresponding DataStructure
(so that we can even do action like go to previous
animation, etc).